requirements.txt
# Core Framework
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38

# LLM Integration
openai==1.30.1  # Optional: only needed if using OpenAI

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1

# HTTP & API
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.5

# Social Media APIs
praw==7.7.1  # Reddit
tweepy==4.14.0  # Twitter/X
snscrape==0.7.0.20230622  # Twitter scraping alternative

# News & RSS
feedparser==6.0.11  # RSS feed parser (FREE, no API keys needed)

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Configuration
python-dotenv==1.0.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Utilities
python-dateutil==2.9.0
pytz==2024.1
colorama==0.4.6  # Colored console output
rich==13.7.1  # Beautiful terminal formatting

# Logging
loguru==0.7.2

# Testing (for development)
pytest==8.2.0
pytest-asyncio==0.23.6

# Rate Limiting
ratelimit==2.2.1

# Data Validation
marshmallow==3.21.1
